requirements.txt
pytest==7.1.3
selenium==4.5.0
webdriver_manager==3.8.3
allure-pytest==2.11.1
pytest-xdist==2.5.0
requests==2.28.1
Faker==15.1.1
Appium-Python-Client==2.7.1
